+++ components/form-builder/app/edit/elements/DropDown.tsx
@@ -1,0 +1,67 @@
+import React from "react";
+import { useSelect, UseSelectStateChange } from "downshift";
+import { ElementOption } from "../../../types";
+import { ChevronDown } from "../../../icons";
+
+export const DropDown = ({
+  items,
+  selectedItem,
+  onChange,
+  ariaLabel,
+}: {
+  items: ElementOption[];
+  selectedItem: ElementOption;
+  onChange: (changes: UseSelectStateChange<ElementOption | null | undefined>) => void;
+  ariaLabel: string;
+}) => {
+  const { isOpen, getToggleButtonProps, getMenuProps, highlightedIndex, getItemProps } = useSelect({
+    items,
+    selectedItem,
+    onSelectedItemChange: onChange,
+  });
+
+  const headerProps = {
+    ...getToggleButtonProps(),
+    "aria-label": ariaLabel,
+    "aria-labelledby": null,
+  };
+
+  return (
+    <div
+      className="builder-element-dropdown"
+      style={{ position: "relative", zIndex: isOpen ? 20000 : 1 }}
+    >
+      <button className="header-button" {...headerProps}>
+        {selectedItem && (
+          <div className="selected" data-testid="element-select-active">
+            {selectedItem.value}
+            <ChevronDown />
+          </div>
+        )}
+      </button>
+      <ul {...getMenuProps()} style={{ display: isOpen ? "block" : "none" }}>
+        {isOpen &&
+          items.map((item, index) => (
+            <li
+              className={highlightedIndex === index ? "highlighted" : ""}
+              key={`${item.id}-${index}`}
+              {...getItemProps({ item, index })}
+            >
+              {item.icon && (
+                <div className={highlightedIndex === index ? "icon icon-highlighted" : "icon"}>
+                  {item.icon}
+                </div>
+              )}
+              <div
+                className={highlightedIndex === index ? "text highlighted" : "text"}
+                data-testid={item.id}
+              >
+                {item.value}
+              </div>
+              {item.append && item.append}
+            </li>
+          ))}
+      </ul>
+    </div>
+  );
+};
