+++ components/form-builder/app/edit/elements/index.ts
@@ -1,6 +1,6 @@
 export { Options } from "./Options";
 export { ShortAnswer } from "./ShortAnswer";
 export { Option } from "./Option";
-export { Select } from "./Select";
+export { DropDown } from "./DropDown";
 export { RichText } from "./RichText";
 export { RichTextLocked } from "./RichTextLocked";
