+++ public/static/locales/en/form-builder.json
@@ -190,5 +190,6 @@
     "phone": "Enter a phone number like, 111-222-3333",
     "date": "Enter a date like, mm/dd/yyyy",
     "number": "Only enter numbers"
-  }
+  },
+  "selectElement": "Select an Element type"
 }
