+++ public/static/locales/fr/form-builder.json
@@ -190,5 +190,6 @@
     "phone": "Entrez un numéro de téléphone comme 111-222-3333",
     "date": "Entrez une date comme mm/dd/yyyy",
     "number": "N'entrez que des nombres"
-  }
+  },
+  "selectElement": "[FR] Select an Element type"
 }
