+++ styles/_form-builder.scss
@@ -124,3 +124,95 @@
     border-top: 1px solid #cacaca;
   }
 }
+
+.builder-element-dropdown {
+  width: 250px;
+
+  .header-button {
+    padding-left: 20px;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    display: flex;
+    background-color: #fff;
+    border-radius: 4px;
+    border: 1.5px solid #000000;
+    width: 100%;
+    line-height: 24px;
+    &:focus,
+    &[aria-expanded="true"] {
+      border-color: #303fc3;
+      box-shadow: 0 0 0 2.5px #303fc3;
+      outline: 0;
+    }
+  }
+
+  .selected {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+
+    svg {
+      margin-right: 20px;
+    }
+  }
+
+  ul {
+    box-shadow: 0 1px 2px 0 rgba(60, 64, 67, 0.3), 0 2px 6px 2px rgba(60, 64, 67, 0.15);
+    padding: 0px;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    border-radius: 4px;
+    margin: 0px;
+    margin-top: 10px;
+    list-style: none;
+    background-color: #fff;
+    position: absolute;
+    width: 250px;
+  }
+
+  li {
+    display: flex;
+    align-content: flex-start;
+    cursor: pointer;
+    padding-left: 20px;
+    padding-right: 26px;
+    padding-bottom: 8px;
+    padding-top: 8px;
+
+    .text {
+      margin-left: 10px;
+      line-height: 24px;
+      color: #202124;
+    }
+    .highlighted {
+      color: #fff;
+    }
+
+    .icon {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      text-align: left;
+      overflow: hidden;
+      display: inline-block;
+      position: relative;
+      padding-top: 2px;
+
+      svg {
+        height: 20px;
+        width: 20px;
+      }
+    }
+    .highlighted {
+      color: #fff;
+    }
+  }
+  .highlighted {
+    background: #718096;
+    svg {
+      path {
+        fill: #fff;
+      }
+    }
+  }
+}
